Restaurant_Backend/requirements.txt
Django==5.2.9
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1